clippy: tidy up allow lints
