+++ src/class/macros.rs
@@ -162,7 +162,6 @@ macro_rules! py_binary_self_func {
             $crate::callback_body!(py, {
                 let slf_ = py.from_borrowed_ptr::<$crate::PyCell<T>>(slf);
                 let arg = py.from_borrowed_ptr::<$crate::PyAny>(arg);
-                #[allow(clippy::needless_question_mark)]
                 call_operator_mut!(py, slf_, $f, arg).convert(py)?;
                 ffi::Py_INCREF(slf);
                 Ok::<_, $crate::err::PyErr>(slf)
@@ -342,7 +341,7 @@ macro_rules! _call_impl {
             $slf, $fn ;
             (match $raw_arg.extract() {
                 Ok(res) => res,
-                _=> return Ok($py.NotImplemented().convert($py)?)
+                _=> return $py.NotImplemented().convert($py)
             })
             $(;$args)*
         )
