+++ src/ffi/cpython/mod.rs
@@ -4,6 +4,7 @@ pub mod abstract_;
 pub mod bytesobject;
 pub mod ceval;
 pub mod code;
+pub mod compile;
 #[cfg(not(PyPy))]
 pub mod dictobject;
 // skipped fileobject.h
@@ -14,6 +15,7 @@ pub mod initconfig;
 // skipped interpreteridobject.h
 pub mod listobject;
 pub mod object;
+pub mod pydebug;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub mod pylifecycle;
 
@@ -22,6 +24,7 @@ pub use self::abstract_::*;
 pub use self::bytesobject::*;
 pub use self::ceval::*;
 pub use self::code::*;
+pub use self::compile::*;
 #[cfg(not(PyPy))]
 pub use self::dictobject::*;
 pub use self::frameobject::*;
@@ -29,5 +32,6 @@ pub use self::frameobject::*;
 pub use self::initconfig::*;
 pub use self::listobject::*;
 pub use self::object::*;
+pub use self::pydebug::*;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub use self::pylifecycle::*;
