+++ src/ffi/datetime.rs
@@ -1,5 +1,3 @@
-#![cfg_attr(feature = "cargo-clippy", allow(clippy::type_complexity))]
-
 //! FFI bindings to the functions and structs defined in `datetime.h`
 //!
 //! This is the unsafe thin  wrapper around the [CPython C API](https://docs.python.org/3/c-api/datetime.html),
