+++ src/ffi/longobject.rs
@@ -1,9 +1,9 @@
 use crate::ffi::object::*;
 use crate::ffi::pyport::Py_ssize_t;
 use libc::size_t;
-use std::os::raw::{
-    c_char, c_double, c_int, c_long, c_longlong, c_uchar, c_ulong, c_ulonglong, c_void,
-};
+#[cfg(not(Py_LIMITED_API))]
+use std::os::raw::c_uchar;
+use std::os::raw::{c_char, c_double, c_int, c_long, c_longlong, c_ulong, c_ulonglong, c_void};
 
 opaque_struct!(PyLongObject);
