+++ src/lib.rs
@@ -170,9 +170,6 @@ mod conversions;
 pub mod derive_utils;
 mod err;
 pub mod exceptions;
-/// Raw ffi declarations for the c interface of python
-#[allow(clippy::upper_case_acronyms)]
-#[allow(clippy::missing_safety_doc)]
 pub mod ffi;
 pub mod freelist;
 mod gil;
