+++ src/types/floatob.rs
@@ -49,7 +49,7 @@ impl IntoPy<PyObject> for f64 {
 
 impl<'source> FromPyObject<'source> for f64 {
     // PyFloat_AsDouble returns -1.0 upon failure
-    #![cfg_attr(feature = "cargo-clippy", allow(clippy::float_cmp))]
+    #![allow(clippy::float_cmp)]
     fn extract(obj: &'source PyAny) -> PyResult<Self> {
         let v = unsafe { ffi::PyFloat_AsDouble(obj.as_ptr()) };
